Create Analysis
This was, in part, a comparative analysis, so the overall approach that I took was to establish a base for comparison using totals and averages, then comparing the percentages of the individual ride type and member type to the baseline.

I then looked at the difference in weekly breakdown by day and the annual breakdown by month to get a feel for when casual and member riders are most active. 

Finally, I checked to see which start/end stations are most commonly used by casual riders and listed the top 15 overlapping stations.